Refactor: move get_file_store to its own factory.py file
Move the factory function from __init__.py to a dedicated factory.py
module for better code organization.
